:recycle: Move summarize attributes to OCEL
