+++ src/ocelescope/src/ocelescope/ocel/util/__init__.py
@@ -1,5 +1,11 @@
+from ocelescope.ocel.util.attributes import summarize_attribute_values
 from ocelescope.ocel.util.hash import hash_string_list
 from ocelescope.ocel.util.relations import RelationCountSummary
 from ocelescope.ocel.util.xes import create_ocel_from_xml
 
-__all__ = ["RelationCountSummary", "hash_string_list", "create_ocel_from_xml"]
+__all__ = [
+    "RelationCountSummary",
+    "hash_string_list",
+    "create_ocel_from_xml",
+    "summarize_attribute_values",
+]
